+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Send email through GC Notify when a new temporary token is generated
 - A UI with tabs on the Form Settings page. [#486](https://github.com/cds-snc/platform-forms-client/pull/573)
 - A tab on the Form Settings page that allows the user to see and refresh the bearer token.
+- A tab on the Form Settings page that allows form access to be enabled / disabled for users.
 - Replace the retrieval Api lambda implementation by an App backEnd API. [#481](https://github.com/cds-snc/platform-forms-client/issues/481)
 
 ### Changed